Synchronize signature hasher engine
Since one of the library intentions is to have a sigle CookieCodec
instance per application context it needs to handle multi thread
scenarios. Unfortunately the SignatureHasher contains a mutable
hasher engine that fully relies on its state, because of this the
synchronize modifier was added to the intrinsict parts where it
is modified.